Fixed 32/64 bit issue

+++ src/plugger.h
@@ -247,7 +247,7 @@ namespace Bu
 
 		void destroy( T *pPlug )
 		{
-			PluginReg *pReg = (PluginReg *)hObj.get((int)pPlug);
+			PluginReg *pReg = (PluginReg *)hObj.get((ptrdiff_t)pPlug);
 			//printf("pReg:  %08X, pPlug: %08X\n", pReg, pPlug );
 			if( pReg == NULL )
 				return;
